Rename `depolarising_noise_model` as `depolarising`
`NoiselessNoiseModel` is merged into `noise_model`.
