+++ graphix/noise_models/depolarising.py
